+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/services.py
@@ -24,5 +24,5 @@ def get_user_by_email(email: str):
     return get_user_model().objects.filter(email__iexact=email).first()
 
 
-def get_user_by_id(user_id):
-    return get_user_model().objects.filter(id=user_id).first()
+def get_active_user_by_id(user_id):
+    return get_user_model().objects.filter(id=user_id, is_active=True).first()
